+++ projects/igniteui-angular/src/lib/grids/grid/grid.crud.spec.ts
@@ -139,7 +139,7 @@ describe('IgxGrid - CRUD operations #grid', () => {
     });
 
     it('should support updating a row through the grid API', () => {
-        spyOn(grid.onRowEdit, 'emit').and.callThrough();
+        spyOn(grid.rowEdit, 'emit').and.callThrough();
         const doneSpy = spyOn(grid.rowEditDone, 'emit').and.callThrough();
 
         // Update non-existing row
@@ -165,7 +165,7 @@ describe('IgxGrid - CRUD operations #grid', () => {
             owner: grid
         };
 
-        (grid as IgxGridComponent).onRowEdit.pipe(first()).subscribe(e => {
+        (grid as IgxGridComponent).rowEdit.pipe(first()).subscribe(e => {
             expect(e).toEqual(args);
             expect(e.rowData).toBe(grid.dataView[0]);
             expect(e.rowData).toBe(e.oldValue);
@@ -174,7 +174,7 @@ describe('IgxGrid - CRUD operations #grid', () => {
         grid.updateRow(args.newValue, 1);
         fix.detectChanges();
 
-        expect(grid.onRowEdit.emit).toHaveBeenCalledTimes(1);
+        expect(grid.rowEdit.emit).toHaveBeenCalledTimes(1);
         expect(grid.rowEditDone.emit).toHaveBeenCalledTimes(1);
         expect(grid.rowEditDone.emit).toHaveBeenCalledWith(doneArgs);
 
@@ -187,7 +187,7 @@ describe('IgxGrid - CRUD operations #grid', () => {
     });
 
     it('should support updating a row through the grid API and event handler', () => {
-        spyOn(grid.onRowEdit, 'emit').and.callThrough();
+        spyOn(grid.rowEdit, 'emit').and.callThrough();
         spyOn(grid.rowEditDone, 'emit').and.callThrough();
 
         const args: IGridEditEventArgs = {
@@ -199,7 +199,7 @@ describe('IgxGrid - CRUD operations #grid', () => {
             owner: grid
         };
 
-        (grid as IgxGridComponent).onRowEdit.pipe(first()).subscribe(e => {
+        (grid as IgxGridComponent).rowEdit.pipe(first()).subscribe(e => {
             expect(e).toEqual(args);
             expect(e.rowData).toBe(grid.dataView[0]);
             expect(e.rowData).toBe(e.oldValue);
@@ -217,17 +217,17 @@ describe('IgxGrid - CRUD operations #grid', () => {
         grid.updateRow(666, 1, 'index');
         fix.detectChanges();
 
-        expect(grid.onRowEdit.emit).toHaveBeenCalledTimes(1);
+        expect(grid.rowEdit.emit).toHaveBeenCalledTimes(1);
         expect(grid.rowEditDone.emit).toHaveBeenCalledTimes(1);
 
-        // expect(grid.onRowEdit.emit).toHaveBeenCalledWith(args);
+        // expect(grid.rowEdit.emit).toHaveBeenCalledWith(args);
         expect(grid.rowEditDone.emit).toHaveBeenCalledWith(handlerUpdateArgs);
         expect(grid.rowList.first.cells.first.value).toEqual(777);
         expect(grid.data[0].index).toEqual(777);
     });
 
     it('should support updating a cell value through the grid API', () => {
-        spyOn(grid.onCellEdit, 'emit').and.callThrough();
+        spyOn(grid.cellEdit, 'emit').and.callThrough();
         const doneSpy = spyOn(grid.cellEditDone, 'emit').and.callThrough();
 
         // Update a non-existing cell
@@ -260,7 +260,7 @@ describe('IgxGrid - CRUD operations #grid', () => {
         expect(grid.rowList.first.cells.first.value).not.toEqual(-100);
         expect(grid.rowList.first.cells.first.nativeElement.textContent).not.toMatch('-100');
 
-        (grid as IgxGridComponent).onCellEdit.pipe(first()).subscribe(e => {
+        (grid as IgxGridComponent).cellEdit.pipe(first()).subscribe(e => {
             expect(e).toEqual(args);
             expect(e.rowData).toBe(grid.dataView[0]);
             expect(e.rowData.index).toBe(e.oldValue);
@@ -269,7 +269,7 @@ describe('IgxGrid - CRUD operations #grid', () => {
         grid.updateCell(200, 1, 'index');
         fix.detectChanges();
 
-        expect(grid.onCellEdit.emit).toHaveBeenCalledTimes(1);
+        expect(grid.cellEdit.emit).toHaveBeenCalledTimes(1);
         expect(grid.cellEditDone.emit).toHaveBeenCalledTimes(1);
         expect(grid.cellEditDone.emit).toHaveBeenCalledWith(doneArgs);
 
@@ -281,7 +281,7 @@ describe('IgxGrid - CRUD operations #grid', () => {
     });
 
     it('should support updating a cell value through the grid API and event handler', () => {
-        spyOn(grid.onCellEdit, 'emit').and.callThrough();
+        spyOn(grid.cellEdit, 'emit').and.callThrough();
         spyOn(grid.cellEditDone, 'emit').and.callThrough();
 
         const cell = grid.getCellByColumn(0, 'index');
@@ -307,7 +307,7 @@ describe('IgxGrid - CRUD operations #grid', () => {
             owner: grid
         };
 
-        (grid as IgxGridComponent).onCellEdit.pipe(first()).subscribe(e => {
+        (grid as IgxGridComponent).cellEdit.pipe(first()).subscribe(e => {
             expect(e).toEqual(args);
         });
 
@@ -493,8 +493,8 @@ describe('IgxGrid - CRUD operations #grid', () => {
             [height]="null"
             (onRowAdded)="rowAdded($event)"
             (onRowDeleted)="rowDeleted($event)"
-            (onCellEdit)="editDone($event)"
-            (onRowEdit)="editDone($event)"
+            (cellEdit)="editDone($event)"
+            (rowEdit)="editDone($event)"
             [autoGenerate]="true"
             [primaryKey]="'index'">
         </igx-grid>